Data-ingestor/requirements.txt
chromadb==1.1.0
langchain==0.3.27
langchain-community==0.3.27
unstructured[all]==0.11.8
python-dotenv
pillow
pytesseract
pdf2image
python-docx>=1.1.0
docx2txt>=0.8
PyPDF2>=3.0.0
python-pptx>=0.6.23
